Initial commit, GUI written by https://github.com/AlveLarsson.
Contains:
- The standard Git files, such as README.md, License, .gitignore...
- The standard NPM files, such as package.json and package-lock.json.
- A svelte/sapper setup.
- The Meros GUI which is currently shimmed with test JSON.
- The exported build files.

This does NOT link with the Meros node at this time, and does need some 
fine tuning. That said, I'm very happy to have this as the bones of this 
project. Once its working with Meros, it'll be further improved in order 
to be the best GUI it can be.